config: move is_config_key_char() to "config.h"
The iskeychar() function in "config.c" checks if a character is valid
for the section or variable name part of a config key.

In a follow up commit we will want to check outside "config.c" if a
string can be a valid variable name of a config key, so will will want
to reuse that fonction.

Let's then move it from "config.c" to "config.h", and, while at it,
let's rename it to is_config_key_char().

Signed-off-by: Christian Couder <chriscool@tuxfamily.org>